backend/requirements.txt
asgiref==3.8.1
Django==4.2.17
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.1.0
PyJWT==2.10.1
pyotp==2.8.0
sqlparse==0.5.3
typing_extensions==4.12.2
qrcode==7.3.1
